sub duty test added

+++ src/main/java/finalproject/finalproject/service/impl/SubDutyServiceImpl.java
@@ -1,11 +1,14 @@
 package finalproject.finalproject.service.impl;
 
 
+import finalproject.finalproject.Entity.duty.Duty;
 import finalproject.finalproject.Entity.duty.SubDuty;
 import finalproject.finalproject.repository.SubDutyRepository;
 import finalproject.finalproject.service.SubDutyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 
 @Service
@@ -17,6 +20,10 @@ public class SubDutyServiceImpl
 
     @Override
     public void deleteSubDutyFromTheExistDuty(SubDuty subDuty) {
+        if (subDuty == null) {
+            throw new IllegalArgumentException("sub duty cannot be null");
+        }
         subDutyRepository.deleteSubDutyFromTheExistDuty(subDuty);
     }
+
 }
